+++ frontend/src/components/projectList.tsx
@@ -1,0 +1,21 @@
+export const ProjectList = () => {
+    return <table>
+        <thead>
+        <tr>
+            <th>No</th>
+            <th>Name</th>
+            <th>Start</th>
+            <th>Ende</th>
+        </tr>
+        </thead>
+        <tbody>
+        <tr>
+            <td>1</td>
+            <td>2</td>
+            <td>3</td>
+            <td>4</td>
+        </tr>
+        </tbody>
+        <tfoot></tfoot>
+    </table>
+}
