+++ frontend/src/components/userInfo.tsx
@@ -1,0 +1,3 @@
+export const UserInfo = () => {
+    return <div>Hallo, Michael</div>
+}
